requirements.txt
googletrans==4.0.0rc1
gTTS==2.2.4
Pillow==9.3.0
requests==2.28.1
SpeechRecognition==3.8.1
streamlit==1.14.1
streamlit_lottie==0.0.3
